+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+11 January 2017 (0.7.0.6)
+
+* Fixed: Settings being not correctly saved in some cases
+* Updated: Translations
 9 January 2017 (0.7.0.5)
 
 * Fixed: Lockfile being not deleted on exit